pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analizer"
authors = ["Abra19 <evabrarova@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Abra19/python-project-83"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
